-added tests for XMLDocumentParser
