+++ tests/ShapeEditorLib.Tests/Files/TestView.xml
