--- tests/ShapeEditorLib.Tests/XMLParserTests.fs
@@ -1,8 +1,0 @@
-﻿module ShapeEditorLibTests
-
-open NUnit.Framework
-open FsUnit
-
-[<Test>]
-let ``some test``() = 
-    1 |> should equal 1
